增加商品详情页

+++ service/src/main/java/com/yisutech/corp/home/service/jfmall/JfMallSrv.java
@@ -52,5 +52,5 @@ public interface JfMallSrv {
 	 * @param productId 商品ID
 	 * @return
 	 */
-	WxExchangeProduct getExchangeProduct(Long productId);
+	WxExchangeProduct getExchangeProduct(Integer productId);
 }
